+++ pinot-controller/src/main/java/org/apache/pinot/controller/util/ServerSegmentMetadataReader.java
@@ -47,7 +47,7 @@ import org.apache.pinot.common.restlet.resources.TableSegments;
 import org.apache.pinot.common.restlet.resources.ValidDocIdsBitmapResponse;
 import org.apache.pinot.common.restlet.resources.ValidDocIdsMetadataInfo;
 import org.apache.pinot.common.utils.RoaringBitmapUtils;
-import org.apache.pinot.segment.local.data.manager.TableDataManager;
+import org.apache.pinot.segment.local.data.manager.NeedRefreshResponse;
 import org.apache.pinot.spi.utils.JsonUtils;
 import org.glassfish.jersey.client.ClientConfig;
 import org.glassfish.jersey.client.ClientProperties;
@@ -398,7 +398,7 @@ public class ServerSegmentMetadataReader {
     return response;
   }
 
-  public Map<String, List<TableDataManager.NeedRefreshResponse>> getSegmentsForRefreshFromServer(
+  public Map<String, List<NeedRefreshResponse>> getSegmentsForRefreshFromServer(
       String tableNameWithType, Set<String> serverInstances, BiMap<String, String> endpoints, int timeoutMs) {
     LOGGER.debug("Getting list of segments for refresh from servers for table {}.", tableNameWithType);
     List<String> serverURLs = new ArrayList<>();
@@ -410,7 +410,7 @@ public class ServerSegmentMetadataReader {
         new CompletionServiceHelper(_executor, _connectionManager, endpointsToServers);
     CompletionServiceHelper.CompletionServiceResponse serviceResponse =
         completionServiceHelper.doMultiGetRequest(serverURLs, tableNameWithType, true, timeoutMs);
-    Map<String, List<TableDataManager.NeedRefreshResponse>> serverResponses = new HashMap<>();
+    Map<String, List<NeedRefreshResponse>> serverResponses = new HashMap<>();
 
     int failedParses = 0;
     for (Map.Entry<String, String> streamResponse : serviceResponse._httpResponses.entrySet()) {
@@ -418,7 +418,7 @@ public class ServerSegmentMetadataReader {
         // TODO: RV - get the instance name instead of the endpoint
         serverResponses.put(streamResponse.getKey(),
             JsonUtils.stringToObject(streamResponse.getValue(),
-                new TypeReference<List<TableDataManager.NeedRefreshResponse>>() { }));
+                new TypeReference<List<NeedRefreshResponse>>() { }));
       } catch (Exception e) {
         failedParses++;
         LOGGER.error("Unable to parse server {} response due to an error: ", streamResponse.getKey(), e);
