+++ pinot-segment-local/src/main/java/org/apache/pinot/segment/local/data/manager/TableDataManager.java
@@ -18,9 +18,6 @@
  */
 package org.apache.pinot.segment.local.data.manager;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.cache.Cache;
 import java.io.File;
@@ -327,40 +324,11 @@ public interface TableDataManager {
   default void onConsumingToOnline(String segmentNameStr) {
   }
 
-  class NeedRefreshResponse {
-    private final String _segmentName;
-    private final boolean _needRefresh;
-    private final String _reason;
-
-    @JsonCreator
-    public NeedRefreshResponse(@JsonProperty("segmentName") String segmentName,
-        @JsonProperty("reason") String reason) {
-      _segmentName = segmentName;
-      _needRefresh = true;
-      _reason = reason;
-    }
-
-    public NeedRefreshResponse(String segmentName, boolean needRefresh, String reason) {
-      _segmentName = segmentName;
-      _needRefresh = needRefresh;
-      _reason = reason;
-    }
-
-    @JsonProperty
-    public String getSegmentName() {
-      return _segmentName;
-    }
-
-    @JsonIgnore
-    public boolean isNeedRefresh() {
-      return _needRefresh;
-    }
-
-    @JsonProperty
-    public String getReason() {
-      return _reason;
-    }
-  }
-
+  /**
+   * Return list of segment names that need to be refreshed along with reason.
+   * @param tableConfig Table Config of the table
+   * @param schema Schema of the table
+   * @return List of {@link NeedRefreshResponse} with segment names and reason for refresh
+   */
   List<NeedRefreshResponse> getSegmentsForRefresh(TableConfig tableConfig, Schema schema);
 }
